Renamed packages to follow conventions
