+++ src/main/java/com/computer_squad/command/class_alarms/CronHelper.java
@@ -1,4 +1,4 @@
-package com.ComputerSquad.commands.ClassAlarm;
+package com.computer_squad.command.class_alarms;
 
 
 import com.cronutils.builder.CronBuilder;
