+++ src/main/java/com/computer_squad/services/UserEvents.java
@@ -1,9 +1,7 @@
-package com.ComputerSquad.jda;
+package com.computer_squad.services;
 
-import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Role;
-import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.guild.member.GuildMemberJoinEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
@@ -18,7 +16,6 @@ public class UserEvents extends ListenerAdapter {
 	public void onGuildMemberJoin(@NotNull GuildMemberJoinEvent event) {
 		// Automatically add a role when an user joins
 		Member member = event.getMember();
-		JDA jda = event.getJDA();
 		if (!roleID.isEmpty()) {
 			// Roles
 			Role role = event.getGuild().getRoleById(roleID);
